Drop a redundant cast the newer clippy rejects
u32be already returns u64 here, so 'l1_size as u64' is a no-op. Local
clippy is 1.94 and CI's is 1.98, which added the lint -- that gap is why a
tree that was clean here failed there.

+++ src/images.rs
@@ -198,7 +198,7 @@ impl Qcow2Reader {
         // than its own L1 table can address. A header claiming 16 exabytes for
         // a 149-byte file had the carver searching a 1.6e19-byte window, which
         // is a hang with no output: cargo-fuzz found exactly that.
-        let addressable = (l1_size as u64)
+        let addressable = l1_size
             .checked_mul(cluster_size / 8)
             .and_then(|clusters| clusters.checked_mul(cluster_size))
             .ok_or_else(|| err("implausible QCOW2 geometry"))?;
